Create a useClickToCopy hook
